split networking.c
